requirements.txt
annotated-types==0.7.0
anyio==4.9.0
awsebcli==3.24.1
bcrypt==4.3.0
blessed==1.21.0
botocore==1.38.23
cement==2.10.14
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
fabric==3.2.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
invoke==2.2.0
jiter==0.10.0
jmespath==1.0.1
openai==1.79.0
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyNaCl==1.5.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
termcolor==2.5.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
certifi
uvicorn[standard]==0.29.0
